+++ src/NugetVersion/GetNugetVersions/GetNugetVersions.cs
@@ -6,20 +6,64 @@ using NuGet.Versioning;
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
+
 namespace NugetVersion.GetNugetVersions
 {
-    internal class NugetPackageVersionUtil
+    public class NugetPackageVersionUtil
     {
         private static ConcurrentDictionary<string, NuGetVersion> _nugetVersionLatest = new();
         private readonly Microsoft.Extensions.Logging.ILogger _logger;
+        private TimeSpan _cacheTimeSpan = TimeSpan.FromMinutes(20);
+
 
         public NugetPackageVersionUtil(Microsoft.Extensions.Logging.ILogger logger)
         {
             _logger = logger;
+        }
+
+        private string GetPathForLocalNugetCache()
+        {
+            return Path.Combine(Path.GetTempPath(), "_nugetversionsCache");
+        }
+
+
+        private string SaveNugetVersionCache(string packageId, NuGetVersion version)
+        {
+            var cachePath = GetPathForLocalNugetCache();
+            if (!Directory.Exists(cachePath))
+            {
+                Directory.CreateDirectory(cachePath);
+            }
+
+            var file = Path.Combine(cachePath, $"{packageId}.json");
+            var content = version.ToFullString();
+            File.WriteAllText(file, content);
+            return file;
+        }
+
+        private NuGetVersion LoadNugetVersionCacheOrNull(string packageId)
+        {
+            var cachePath = GetPathForLocalNugetCache();
+            var file = Path.Combine(cachePath, $"{packageId}.json");
+            if (!File.Exists(file))
+            {
+                return null;
+            }
+
+            // has it expired
+            if (new FileInfo(file).CreationTime.Add(_cacheTimeSpan) < DateTime.Now)
+            {
+                return null;
+            }
+
+            var fullVersionString = File.ReadAllText(file);
+            var model = new NuGetVersion(fullVersionString);
+            return model;
         }
 
         public async Task<NuGetVersion> GetLatestNugetPackageVersionAsync(string packageId)
@@ -28,9 +72,21 @@ namespace NugetVersion.GetNugetVersions
             {
                 try
                 {
-                    var versions = await GetNugetVersionsForPackageAsync(packageId);
-                    var version = versions.OrderByDescending(x => x.Version)
-                        .FirstOrDefault(x => !x.IsPrerelease);
+                    // 1. try file cache
+                    NuGetVersion version = LoadNugetVersionCacheOrNull(packageId);
+                    if (version == null)
+                    {
+                        // 2. try remote source
+                        var versions = await GetNugetVersionsForPackageAsync(packageId);
+                        version = versions.OrderByDescending(x => x.Version)
+                            .FirstOrDefault(x => !x.IsPrerelease);
+
+                        if (version != null)
+                        {
+                            // 3. save to cache
+                            SaveNugetVersionCache(packageId, version);
+                        }
+                    }
 
                     _nugetVersionLatest.AddOrUpdate(packageId, version, (key, oldValue) => version);
                     _logger.LogInformation("Package: {PackageId} latest version {version} added to cache", packageId, version);