+++ src/NugetVersion/Renderer/IProjectFileResultsRenderer.cs
@@ -1,12 +1,14 @@
-using System.Collections.Generic;
+using NuGet.Versioning;
 using NugetVersion.Models;
 using NugetVersion.Project;
+using System.Collections.Generic;
 
 namespace NugetVersion.Renderer
 {
     public interface IProjectFileResultsRenderer
     {
         OutputFileFormat Format { get; }
-        void RenderResults(string basePath, SearchQueryFilter filter, IEnumerable<ProjectFile> projFiles);
+        void RenderResults(string basePath, SearchQueryFilter filter, IEnumerable<ProjectFile> projFiles,
+            IDictionary<string, NuGetVersion> latestPackageVersions);
     }
 }